Added finalization functionality

+++ app/src/main/java/com/its/vdv/rest/request/AddLikeRequest.java
@@ -7,5 +7,5 @@ import lombok.Setter;
 @Setter
 public class AddLikeRequest {
     private long vdvid;
-    private int weight = 1;
+    private long weight = 1;
 }
